remove some redundant yaml file parser using inline funcs
